API-Treino-de-Corrida/requirements.txt
psycopg2-binary
python-dotenv 
tabulate